TabTray Delegate Separation And Sync View Placement
