+++ Client/Frontend/Browser/Tab Tray/Views/TabTrayCell.swift
