+++ Client/Frontend/Browser/Tab Tray/Views/TabTrayPrivateModeInfoView.swift
